# perf: store/cachekv: preallocate kvL in dirtyItems which gets appended too