Noticed in RAM profiles that the append for kvL had a concentrated amount of RAM consumption yet where we invoke make didn't have any capacity. Alas on profiling the results show a distribution which eases on allocation pressure and distributes allocations across different sections:

* Before:
```shell
         .          .    342:	kvL := make([]*kv.Pair, 0)
         .          .    343:	for i := startIndex; i <= endIndex; i++ {
         .          .    344:		key := strL[i]
         .          .    345:		cacheValue := store.cache[key]
    1.14GB     1.14GB    346:		kvL = append(kvL, &kv.Pair{Key: []byte(key), Value: cacheValue.value})
         .          .    347:	}
```

* After:
```shell
   76.33MB    76.33MB    342:	kvL := make([]*kv.Pair, 0, endIndex-startIndex)
         .          .    343:	for i := startIndex; i <= endIndex; i++ {
         .          .    344:		key := strL[i]
         .          .    345:		cacheValue := store.cache[key]
  851.42MB   851.42MB    346:		kvL = append(kvL, &kv.Pair{Key: []byte(key), Value: cacheValue.value})
         .          .    347:	}
```

which also shows up in benchmark results:

```shell
$ benchstat before.txt after.txt
name                             old time/op    new time/op    delta
IteratorOnParentWith1MDeletes-8     1.69s ± 5%     1.65s ± 6%    ~     (p=0.095 n=9+10)

name                             old alloc/op   new alloc/op   delta
IteratorOnParentWith1MDeletes-8     238MB ± 0%     214MB ± 0%  -9.94%  (p=0.000 n=10+9)

name                             old allocs/op  new allocs/op  delta
IteratorOnParentWith1MDeletes-8     3.10M ± 0%     3.10M ± 0%  -0.00%  (p=0.000 n=10+10)
```

Fixes #14167